refactor(definitions): 🔨 split event enums into dedicated files
