+++ Sources/EventViewerX/Definitions/BitLockerVolumeType.cs
@@ -1,0 +1,10 @@
+namespace EventViewerX;
+
+/// <summary>
+/// Types of volumes protected by BitLocker.
+/// </summary>
+public enum BitLockerVolumeType {
+    OperatingSystemVolume = 810,
+    FixedDataVolume = 811,
+    RemovableDataVolume = 812
+}
