+++ Sources/EventViewerX/Definitions/ElevatedToken.cs
@@ -1,0 +1,9 @@
+namespace EventViewerX;
+
+/// <summary>
+/// Specifies if an elevated token was used for the logon.
+/// </summary>
+public enum ElevatedToken {
+    Yes = 1842,
+    No = 1843
+}
