+++ Sources/EventViewerX/Definitions/PreAuthType.cs
@@ -1,0 +1,8 @@
+namespace EventViewerX;
+
+/// <summary>
+/// Represents Kerberos pre-authentication types.
+/// </summary>
+public enum PreAuthType {
+    // Add appropriate values here
+}
