+++ Sources/EventViewerX/Definitions/Status.cs
@@ -1,0 +1,8 @@
+namespace EventViewerX;
+
+/// <summary>
+/// General status enumeration placeholder.
+/// </summary>
+public enum Status {
+    // Add appropriate values here
+}
