+++ Sources/EventViewerX/Definitions/TicketOptions.cs
@@ -1,0 +1,8 @@
+namespace EventViewerX;
+
+/// <summary>
+/// Flags used to describe Kerberos ticket options.
+/// </summary>
+public enum TicketOptions {
+    // Add appropriate values here
+}
