+++ Sources/EventViewerX/Definitions/VirtualAccount.cs
@@ -1,0 +1,9 @@
+namespace EventViewerX;
+
+/// <summary>
+/// Indicates whether a virtual account was used during logon.
+/// </summary>
+public enum VirtualAccount {
+    Yes = 1843,
+    No = 1844
+}
